Implement Rendererable typeclass
This implements the Renderable type class, which the Game and Level implement to
produce a renderable game.

The Renderable type class requires either "getSprite" or "getSprites" function
to be implemented. Each implementation produces Sprite(s) based on the object
passed. The Sprite(s) are a simple (Point, String) tuple for now.

In doing all this, the Renderer depends on the Game state no longer! The draw
function can be as simple as "render game display".