Implement UI using concrete data type
The typeclasses were pointing towards an antipattern. The UI data type
is used so that we can have a specification for each different UI
implementation. This is also better than typeclasses since we can have
more encapsulation (and don't need to leak things out like the Window
from the underlying library).

We might want to look back at typeclasses when we have an abstracted Map
type for the engine. This way we could have something like

class CanRender map where
    render :: map -> IO ()

And implement render based on the chosen rendering target. However, this
has the disadvantage of not being able to change the renderer during
runtime (such as via CLI arguments, config, etc.). That would require a
more specialized monad which increases complexity. KISS for now.